moved h5 reading into orchestrator for efficiency
